+++ xrlint/plugins/core/__init__.py
@@ -16,6 +16,7 @@ def export_plugin() -> Plugin:
                 "dataset-title-attr": "warn",
                 "grid-mappings": "error",
                 "no-empty-attrs": "warn",
+                "time-coords": "error",
                 "var-units-attr": "warn",
             },
         }